+++ CHANGES.rst
@@ -4,8 +4,8 @@ Changelog
 2.5.1rc1 "Cactus cat" (2024-08-09)
 ----------------------------------
 
--   Enable SSM alerts and HasSSM source property.
-    This adds !1473.
+-   Update trials factor to account for MLy and SPIIR joining O4b.
+
 
 2.5.0 "Cactus cat" (2024-08-08)
 -------------------------------